Create understandingThe Output
The program outputs two kinds of data. 
Firstly, the live stream from the camera with the points marked(if a prerecorded video needs to be used instead of live stream, set the video capture value to the absolute path of the file instead of 0). The live stream window can be closed by pressing the key 'q'. 
Secondly, the program also has an output channel printing the label and the location of the data points. In case any object is not found the program will return a null nested list.